% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_build_url_classification_cerrado}
\alias{aws_build_url_classification_cerrado}
\title{Build AWS URL path for Cerrado}
\usage{
aws_build_url_classification_cerrado(version, year)
}
\arguments{
\item{version}{Character with the version label used in the output path (e.g. \code{"v1"}).}

\item{year}{Integer or character with the year used in the output path and filename.}
}
\value{
Character with URL classification file.
}
\description{
Builds AWS URL for Cerrado in Restore-plus S3
}
\author{
Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Felipe Carlos, \email{efelipecarlos@gmail.com}
}
