% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_colors.R
\name{colors_amazon_mcti}
\alias{colors_amazon_mcti}
\title{Colors Amazon MCTI}
\usage{
colors_amazon_mcti()
}
\value{
A tibble with the colors for the Amazon maps produced for the MCTI.
}
\description{
Returns the colors for the Amazon maps produced for the MCTI.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
