% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_utils.R
\name{contextual_cleaner}
\alias{contextual_cleaner}
\title{Clean contextual NA values in a cube}
\usage{
contextual_cleaner(
  cube,
  window_size = 3L,
  target_class = 1L,
  mode_class = 2L,
  memsize = 4L,
  multicores = 2L,
  output_dir,
  version = "v1",
  progress = TRUE
)
}
\arguments{
\item{cube}{A \code{sits} cube.}

\item{window_size}{Integer representing the window size.}

\item{target_class}{Integer representing the target class.}

\item{mode_class}{Integer representing the mode class.}

\item{memsize}{Integer representing the memory size.}

\item{multicores}{Integer representing the number of cores to use.}

\item{output_dir}{Character representing the output directory.}

\item{version}{Character representing the version of the cube.}

\item{progress}{Logical representing whether to show progress.}
}
\value{
A \code{sits} cube with the cleaned contextual NA values.
}
\description{
Cleans contextual NA values in a cube.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
