% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{create_data_dir}
\alias{create_data_dir}
\title{Create a data directory}
\usage{
create_data_dir(base_dir, name)
}
\value{
Character representing the data directory.
}
\description{
Creates a data directory.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
