% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_utils.R
\name{crop_to_tiles}
\alias{crop_to_tiles}
\title{Crop a cube to a set of tiles}
\usage{
crop_to_tiles(
  cube,
  tiles,
  output_dir,
  multicores = 10,
  grid_system = "BDC_MD_V2"
)
}
\arguments{
\item{cube}{A \code{sits} cube.}

\item{tiles}{A character vector of tile names.}

\item{output_dir}{Character path for intermediate files.}

\item{multicores}{Integer specifying the number of cores for parallel
processing (default: 10).}

\item{grid_system}{Character specifying the grid system to use.
(default: "BDC_MD_V2").}
}
\value{
A character vector with the paths to the cropped tile files.
}
\description{
Crops a cube to a set of tiles.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
