% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{crs_bdc}
\alias{crs_bdc}
\title{Get Brazil Data Cube (BDC) coordinate reference system}
\usage{
crs_bdc()
}
\value{
Character representing the Brazil Data Cube CRS
}
\description{
Returns the CRS object of the Brazil Data Cube (BDC) data.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
