% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{cube_generate_indices_bdc}
\alias{cube_generate_indices_bdc}
\title{Generate indices for a BDC cube}
\usage{
cube_generate_indices_bdc(cube, output_dir, multicores, memsize)
}
\value{
A sits cube with the generated indices.
}
\description{
Generates indices for a BDC cube.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
