% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{cube_load}
\alias{cube_load}
\title{Load a cube (patching STAC address)}
\usage{
cube_load(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link[sits]{sits_cube}}.}
}
\value{
A sits cube.
}
\description{
Loads a cube and patches the STAC address if the STAC
address (\code{RESTORE_PLUS_STAC_ADDRESS}) environment
variable is set.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
