% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap.R
\name{cube_remap}
\alias{cube_remap}
\title{Remap cube labels}
\usage{
cube_remap(
  cube,
  output_dir,
  multicores = 10,
  memsize = 16,
  mapping_reference,
  rules = NULL
)
}
\arguments{
\item{cube}{A \code{sits} cube.}

\item{output_dir}{Character specifying the output directory.}

\item{multicores}{Integer specifying the number of cores for parallel
processing (default: 10).}

\item{memsize}{Integer specifying the memory size for parallel
processing (default: 16).}

\item{mapping_reference}{A mapping reference object.}

\item{rules}{A rules object.}
}
\value{
A \code{sits} cube with the remapped labels.
}
\description{
Remaps the labels of a cube.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
