% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_utils.R
\name{detect_chunks_with_na}
\alias{detect_chunks_with_na}
\title{Detect chunks with NA values}
\usage{
detect_chunks_with_na(files, version, multicores, memsize, output_dir)
}
\arguments{
\item{files}{Character representing the input raster files.}

\item{version}{Character representing the version of the files.}

\item{multicores}{Integer representing the number of cores to use.}

\item{memsize}{Integer representing the memory size.}

\item{output_dir}{Character representing the output directory.}
}
\value{
Character representing the output raster file.
}
\description{
Detects chunks with NA values.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
