% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_utils.R
\name{.contextual_cleaner_tile}
\alias{.contextual_cleaner_tile}
\title{Clean contextual NA values in a tile}
\usage{
.contextual_cleaner_tile(
  tile,
  block,
  band,
  window_size,
  target_class,
  mode_class,
  overlap,
  output_dir,
  version,
  progress
)
}
\arguments{
\item{tile}{A \code{sits} tile.}

\item{block}{A \code{sits} block.}

\item{band}{Character representing the band to clean.}

\item{window_size}{Integer representing the window size.}

\item{target_class}{Integer representing the target class.}

\item{mode_class}{Integer representing the mode class.}

\item{overlap}{Integer representing the overlap.}

\item{output_dir}{Character representing the output directory.}

\item{version}{Character representing the version of the tile.}

\item{progress}{Logical representing whether to show progress.}
}
\value{
A \code{sits} tile with the cleaned contextual NA values.
}
\description{
Cleans contextual NA values in a tile.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
