% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{.cube_stac_address}
\alias{.cube_stac_address}
\title{Get the STAC address}
\usage{
.cube_stac_address()
}
\value{
Character representing the STAC address.
}
\description{
Gets the STAC address from the environment variables.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
