% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{.flux_build_experiment_body}
\alias{.flux_build_experiment_body}
\title{Build experiment body}
\usage{
.flux_build_experiment_body(config)
}
\arguments{
\item{config}{Config object}
}
\value{
List with experiment body.
}
\description{
Build required body to create an experiment in Flux
}
\keyword{internal}
