% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_utils.R
\name{.grass_find_path}
\alias{.grass_find_path}
\title{Find GRASS GIS installation path}
\usage{
.grass_find_path()
}
\value{
Character with the GRASS GIS installation path.
}
\description{
Detects the GRASS GIS installation path. First checks the
\code{GRASS_PATH} environment variable, then tries to detect
it from the system using \code{grass --config path}.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
