% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_utils.R
\name{.na_cleaner}
\alias{.na_cleaner}
\title{Clean NA values in a tile}
\usage{
.na_cleaner(
  tile,
  block,
  band,
  roi,
  window_size,
  overlap,
  output_dir,
  version,
  progress
)
}
\arguments{
\item{tile}{A \code{sits} tile.}

\item{block}{A \code{sits} block.}

\item{band}{Character representing the band to clean.}

\item{roi}{A \code{sits} ROI.}

\item{window_size}{Integer representing the window size.}

\item{overlap}{Integer representing the overlap.}

\item{output_dir}{Character representing the output directory.}

\item{version}{Character representing the version of the tile.}

\item{progress}{Logical representing whether to show progress.}
}
\value{
A \code{sits} tile with the cleaned NA values.
}
\description{
Cleans NA values in a tile.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
