% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qml_utils.R
\name{.qml_to_tibble}
\alias{.qml_to_tibble}
\title{Read QML file}
\usage{
.qml_to_tibble(qml_file)
}
\arguments{
\item{qml_file}{Path to QML file}
}
\value{
A tibble with the pixel values, colors, and labels.
}
\description{
Reads a QML file and returns a tibble with the pixel values,
colors, and labels.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
