% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify_utils.R
\name{.reclassify_temporal_consistency_reference}
\alias{.reclassify_temporal_consistency_reference}
\title{Reclassify temporal consistency using reference class}
\usage{
.reclassify_temporal_consistency_reference(
  files,
  reference_id,
  target_id,
  version,
  multicores,
  memsize,
  output_dir
)
}
\arguments{
\item{files}{Character vector of input files}

\item{reference_id}{Integer representing the reference class id}

\item{target_id}{Integer representing the target class id}

\item{version}{Character representing the version}

\item{multicores}{Integer representing the number of cores}

\item{memsize}{Integer representing the memory size}

\item{output_dir}{Character representing the output directory}
}
\value{
Character representing the output file
}
\description{
The function checks over time whether a class is
present and, if so, reclassifies previous years
to the target class. This temporal consistency is useful for
classes that can not be presented in the future if they are not
present in the past (e.g., forest in INPE-PRODES maps, we can't
have something else in the past and forest in the future as we are
not creating original forest).
}
\author{
Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Felipe Carlos, \email{efelipecarlos@gmail.com}
}
\keyword{internal}
