% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi_cerrado.R
\name{.roi_cerrado_ecoregion_name}
\alias{.roi_cerrado_ecoregion_name}
\title{Get the name of the Cerrado ecoregion}
\usage{
.roi_cerrado_ecoregion_name(region_id)
}
\arguments{
\item{region_id}{Integer with the region id (1 to 4).}
}
\value{
Character name of the Cerrado ecoregion
}
\description{
This function returns the name of the Cerrado ecoregion
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
