% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi_cerrado.R
\name{.roi_cerrado_ecoregion_sf}
\alias{.roi_cerrado_ecoregion_sf}
\title{Get the sf object for the Cerrado ecoregions}
\usage{
.roi_cerrado_ecoregion_sf(crs, region_id = NULL)
}
\arguments{
\item{crs}{Character with the CRS to use. If not provided, the CRS of the data file is used.}

\item{region_id}{Integer with the region id (1 to 4). If not provided, all regions are returned.}
}
\value{
sf object for the Cerrado ecoregions
}
\description{
This function returns the sf object for the Cerrado ecoregions.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
