% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_merge_aws.R
\name{.s3_gdal_download_s3_files}
\alias{.s3_gdal_download_s3_files}
\title{Download S3 files to local directory}
\usage{
.s3_gdal_download_s3_files(bucket, keys, dest_dir)
}
\arguments{
\item{bucket}{S3 bucket name}

\item{keys}{Character vector of S3 keys to download}

\item{dest_dir}{Local directory to download files to}
}
\value{
Character vector of local file paths
}
\description{
Download S3 files to local directory
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
