% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_merge_aws.R
\name{.s3_gdal_get_s3_keys}
\alias{.s3_gdal_get_s3_keys}
\title{Get list of S3 keys matching pattern}
\usage{
.s3_gdal_get_s3_keys(bucket, prefix, pattern, recursive = FALSE)
}
\arguments{
\item{bucket}{S3 bucket name}

\item{prefix}{Key prefix to search within}

\item{pattern}{File pattern to match (regex style)}

\item{recursive}{Logical indicating whether to search recursively}
}
\value{
Character vector of matching S3 keys
}
\description{
Get list of S3 keys matching pattern
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
