% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_merge_aws.R
\name{.s3_gdal_is_s3_path}
\alias{.s3_gdal_is_s3_path}
\title{Check if path is an S3 path}
\usage{
.s3_gdal_is_s3_path(path)
}
\arguments{
\item{path}{Character string representing a file path}
}
\value{
Logical indicating whether the path is an S3 path
}
\description{
Check if path is an S3 path
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
