% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_merge_aws.R
\name{.s3_gdal_merge_rasters}
\alias{.s3_gdal_merge_rasters}
\title{Merge raster files using GDAL}
\usage{
.s3_gdal_merge_rasters(input_files, output_file, type, roi = NULL)
}
\arguments{
\item{input_files}{Character vector of input raster file paths}

\item{output_file}{Output file path for merged raster}

\item{type}{Output raster type (GDAL Names - e.g., Bytes)}

\item{roi}{Optional region of interest (sf object or file path)}
}
\value{
Path to the merged file
}
\description{
Merge raster files using GDAL
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
