% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_merge_aws.R
\name{.s3_gdal_parse_s3_path}
\alias{.s3_gdal_parse_s3_path}
\title{Parse S3 path into bucket and key}
\usage{
.s3_gdal_parse_s3_path(s3_path)
}
\arguments{
\item{s3_path}{Character string with S3 path (s3://bucket/key)}
}
\value{
Named list with bucket and key components
}
\description{
Parse S3 path into bucket and key
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
