% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_merge_aws.R
\name{.s3_gdal_process_roi}
\alias{.s3_gdal_process_roi}
\title{Process ROI input into file path}
\usage{
.s3_gdal_process_roi(roi)
}
\arguments{
\item{roi}{Region of interest (sf object or file path)}
}
\value{
File path to ROI
}
\description{
Process ROI input into file path
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
