% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_merge_aws.R
\name{.s3_gdal_remove_pattern}
\alias{.s3_gdal_remove_pattern}
\title{Remove pattern from filename}
\usage{
.s3_gdal_remove_pattern(filename, pattern)
}
\arguments{
\item{filename}{Character string with filename}

\item{pattern}{Pattern to remove from filename}
}
\value{
Character string with pattern removed
}
\description{
Remove pattern from filename
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
