% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_merge_aws.R
\name{.s3_gdal_upload_to_s3}
\alias{.s3_gdal_upload_to_s3}
\title{Upload file to S3}
\usage{
.s3_gdal_upload_to_s3(local_file, bucket, key)
}
\arguments{
\item{local_file}{Local file path to upload}

\item{bucket}{S3 bucket name}

\item{key}{S3 key (path) for uploaded file}
}
\value{
S3 path to uploaded file
}
\description{
Upload file to S3
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
