% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{flux_experiment}
\alias{flux_experiment}
\title{Create or register a Flux experiment}
\usage{
flux_experiment(config_file = ".restore.json")
}
\arguments{
\item{config_file}{Path to .restore.json file (defaults to ".restore.json")}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       List with experiment details (id, status, context, settings)
}\if{html}{\out{</div>}}
}
\description{
Register a new experiment in the Flux system.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}
}
