% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_utils.R
\name{gdal_addo}
\alias{gdal_addo}
\title{Add overviews into tiff files}
\usage{
gdal_addo(cube)
}
\arguments{
\item{cube}{a sits classification cube.}
}
\value{
character with classified files.
}
\description{
Add overviews into tiff files
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
