% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_utils.R
\name{geom_adjust_map_variations}
\alias{geom_adjust_map_variations}
\title{Adjust map geometry variations}
\usage{
geom_adjust_map_variations(
  file_adjust,
  file_reference,
  year,
  version,
  multicores,
  memsize,
  output_dir
)
}
\arguments{
\item{file_adjust}{Character representing the file to adjust.}

\item{file_reference}{Character representing the reference file.}

\item{year}{Integer representing the year of the map.}

\item{version}{Character representing the version of the map.}

\item{multicores}{Integer representing the number of cores to use.}

\item{memsize}{Integer representing the memory size to use.}

\item{output_dir}{Character representing the output directory.}
}
\value{
Character representing the adjusted map.
}
\description{
Adjusts the geometry variations of a map using a reference map.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
