% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_mask_file_year}
\alias{get_mask_file_year}
\title{Get the year from a mask file}
\usage{
get_mask_file_year(file)
}
\value{
Integer representing the year.
}
\description{
Gets the year from a mask file.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
