% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_utils.R
\name{grass_batch_rreport_generator}
\alias{grass_batch_rreport_generator}
\title{Generate GRASS GIS batch r.report file}
\usage{
grass_batch_rreport_generator(files, output_dir)
}
\arguments{
\item{files}{Character vector specifying the files to process}

\item{output_dir}{Character specifying the output directory}
}
\value{
Path to the generated batch file
}
\description{
Generates a GRASS GIS batch r.report file.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
