% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_utils.R
\name{grass_default_naming_fnc}
\alias{grass_default_naming_fnc}
\title{Generate a default naming function for GRASS GIS r.report output files}
\usage{
grass_default_naming_fnc(file_name)
}
\value{
A function that generates a default naming function for GRASS GIS r.report output files.
}
\description{
Generates a default naming function for GRASS GIS r.report output files.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
