% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_utils.R
\name{grass_parse_rreport}
\alias{grass_parse_rreport}
\title{Parse GRASS GIS r.report output file}
\usage{
grass_parse_rreport(path)
}
\arguments{
\item{path}{Character specifying the file path to the r.report output}
}
\value{
A tibble with columns \code{category} and \code{area_m2}.
}
\description{
Parses a GRASS GIS r.report output file.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
