% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_utils.R
\name{grass_rreport_to_excel}
\alias{grass_rreport_to_excel}
\title{Process GRASS r.report files with QML labels}
\usage{
grass_rreport_to_excel(files)
}
\arguments{
\item{files}{Character vector specifying the r.report file paths to process}
}
\value{
A list of data frames with category, label, and area information
}
\description{
Processes GRASS GIS r.report files by merging with QML label data
and exporting to Excel format. This function assumes a QML file exists in the
same directory as each r.report file.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
