% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_labels.R
\name{labels_amazon_mcti}
\alias{labels_amazon_mcti}
\title{Labels Amazon MCTI}
\usage{
labels_amazon_mcti()
}
\value{
A vector of labels.
}
\description{
Returns the labels for the Amazon maps produced for the MCTI.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
