% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_labels.R
\name{labels_cerrado_classification}
\alias{labels_cerrado_classification}
\title{Labels Cerrado of a single classification}
\usage{
labels_cerrado_classification()
}
\value{
A vector of labels.
}
\description{
Returns the labels for the Cerrado classification.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
