% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{load_restore_mosaic}
\alias{load_restore_mosaic}
\title{Load Restore Mosaic}
\usage{
load_restore_mosaic(data_dir, multicores = 32, memsize = 120, labels, ...)
}
\arguments{
\item{data_dir}{The data directory.}

\item{multicores}{The number of cores to use.}

\item{memsize}{The memory size to use.}

\item{labels}{The labels to use. Default is NULL.}

\item{...}{Additional arguments to pass to the sits_cube function.}
}
\value{
A sits cube.
}
\description{
Loads a restore mosaic from a data directory.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
