% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic.R
\name{mosaic_cube_glad}
\alias{mosaic_cube_glad}
\title{Mosaic a cube}
\usage{
mosaic_cube_glad(
  cube,
  output_dir,
  multicores = 64,
  bands = NULL,
  roi_file = NULL,
  mbtiles = FALSE
)
}
\arguments{
\item{cube}{A \code{sits} cube.}

\item{output_dir}{Character path for intermediate files.}

\item{multicores}{Integer specifying the number of cores for parallel
processing (default: 64).}

\item{bands}{Character vector of band names to include in the mosaic.
Default is NULL, which includes all bands.}

\item{roi_file}{Optional region of interest (ROI) to filter chunks.
Default is NULL.}

\item{mbtiles}{Logical indicating whether to create MBTILES files.
Default is FALSE.}
}
\value{
A tibble with the mosaic files.
}
\description{
Mosaic a cube.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
