% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_utils.R
\name{na_cleaner}
\alias{na_cleaner}
\title{Clean NA values in a cube}
\usage{
na_cleaner(
  cube,
  window_size = 3L,
  memsize = 4L,
  multicores = 2L,
  output_dir,
  version = "v1",
  roi = NULL,
  progress = TRUE
)
}
\arguments{
\item{cube}{A \code{sits} cube.}

\item{window_size}{Integer representing the window size.}

\item{memsize}{Integer representing the memory size.}

\item{multicores}{Integer representing the number of cores to use.}

\item{output_dir}{Character representing the output directory.}

\item{version}{Character representing the version of the cube.}

\item{roi}{A \code{sits} ROI.}

\item{progress}{Logical representing whether to show progress.}
}
\value{
A \code{sits} cube with the cleaned NA values.
}
\description{
Cleans NA values in a cube.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
