% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify_utils.R
\name{notify}
\alias{notify}
\title{Send a notification (using ntfy)}
\usage{
notify(context, message)
}
\arguments{
\item{context}{Character representing the context.}

\item{message}{Character representing the message.}
}
\value{
Nothing.
}
\description{
Sends a notification using ntfy.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
