% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_cubes_dir}
\alias{project_cubes_dir}
\title{Get the Directory path for Cubes}
\usage{
project_cubes_dir()
}
\value{
A character string representing the directory path, constructed
using \code{fs::path()}.
}
\description{
This function retrieves the directory path where cube is stored.
}
\details{
It first checks for an environment variable (\code{RESTORE_PLUS_REG3_DIR})
and falls back to a default directory (\code{data/derived/cubes}) if the
environment variable is not set. The function uses \code{fs::path()} to
ensure proper path construction across different operating systems.
}
