% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_rat.R
\name{rat_set_style}
\alias{rat_set_style}
\title{Configure QML style in a raster as RAT}
\usage{
rat_set_style(map, qml, band = 1)
}
\arguments{
\item{map}{Path to raster file}

\item{qml}{Path to QML file}

\item{band}{Band number}
}
\value{
Path to raster file
}
\description{
Configures the QML style in a raster using RAT.
}
\keyword{internal}
