% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify_utils.R
\name{reclassify_remap_pixels}
\alias{reclassify_remap_pixels}
\title{Remap pixels}
\usage{
reclassify_remap_pixels(file, file_out, rules, multicores, memsize, output_dir)
}
\arguments{
\item{file}{Character representing the input file}

\item{file_out}{Character representing the output file}

\item{rules}{tibble with columns \code{source} and \code{target} representing the
remapping rules}

\item{multicores}{Integer representing the number of cores}

\item{memsize}{Integer representing the memory size}

\item{output_dir}{Character representing the output directory}
}
\value{
Character representing the output file
}
\description{
The function remaps the pixels of a raster file from a
source to a target based on a set of rules.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
