% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify_utils.R
\name{reclassify_temporal_results_to_maps}
\alias{reclassify_temporal_results_to_maps}
\title{Extract temporal results in brick format to individual maps}
\usage{
reclassify_temporal_results_to_maps(
  years,
  output_dir,
  file_brick,
  version = "v1"
)
}
\arguments{
\item{years}{Integer vector of years}

\item{output_dir}{Character representing the output directory}

\item{file_brick}{Character representing the input brick file}

\item{version}{Character representing the version of the classification}
}
\value{
Character vector of output files
}
\description{
The function extracts the temporal results in brick format to individual maps by year.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
