% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_utils.R
\name{replace_na}
\alias{replace_na}
\title{Replace NA values by a value in a raster file}
\usage{
replace_na(file, year, replace_value, version, multicores, memsize, output_dir)
}
\arguments{
\item{file}{Character representing the input raster file.}

\item{year}{Integer representing the year of the raster file.}

\item{replace_value}{Integer representing the value to replace NA values.}

\item{version}{Character representing the version of the raster file.}

\item{multicores}{Integer representing the number of cores to use.}

\item{memsize}{Integer representing the memory size to use.}

\item{output_dir}{Character representing the output directory.}
}
\value{
Character representing the output raster file.
}
\description{
Replaces NA values by a value in a raster file.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
