% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi_cerrado.R
\name{roi_cerrado_regions}
\alias{roi_cerrado_regions}
\title{Get the sf object for the Cerrado ecoregions}
\usage{
roi_cerrado_regions(
  region_id,
  crs = NULL,
  as_file = FALSE,
  as_union = FALSE,
  as_convex = FALSE
)
}
\arguments{
\item{region_id}{Integer with the region id (1 to 4).}

\item{crs}{Character with the CRS to use. If not provided, the CRS of the data file is used.}

\item{as_file}{Logical with whether to return the \code{sf} object as a file path.}

\item{as_union}{Logical with whether to union the \code{sf} objects.}

\item{as_convex}{Logical with whether to return the \code{sf} object as a convex hull.}
}
\value{
\code{sf} or file path object for the Cerrado ecoregions
}
\description{
This function returns the sf object for the Cerrado ecoregions.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
