% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_utils.R
\name{self_update}
\alias{self_update}
\title{Update the package from GitHub}
\usage{
self_update()
}
\value{
Nothing.
}
\description{
Updates the package from GitHub.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
