% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi_cerrado.R
\name{tiles_cerrado_biome}
\alias{tiles_cerrado_biome}
\title{Get the tiles for the Cerrado biome}
\usage{
tiles_cerrado_biome(grid_system = "BDC_MD_V2")
}
\arguments{
\item{grid_system}{Character with the grid system to use.}
}
\value{
\code{sf} object with the tiles for the Cerrado biome
}
\description{
This function returns the tiles for the Cerrado biome.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
