% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{upload_aws}
\alias{upload_aws}
\title{Upload files to AWS S3 bucket}
\usage{
upload_aws(files, object, bucket, ...)
}
\arguments{
\item{files}{Character vector of local file paths to upload}

\item{object}{Character vector specifying S3 object path prefix(es). Can be a single
value (used for all files) or same length as files (one per file)}

\item{bucket}{Character specifying the S3 bucket name}

\item{...}{Additional parameters passed to \code{aws.s3::put_object()}}
}
\value{
A list of results from \code{aws.s3::put_object()} for each uploaded file
}
\description{
Uploads one or more local files to an AWS S3 bucket.
}
\author{
Felipe Carvalho, \email{felipe.carvalho@inpe.br}

Felipe Carlos, \email{efelipecarlos@gmail.com}
}
