% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_versions.R
\name{validation_compare_versions}
\alias{validation_compare_versions}
\title{Compare versions of a classification cube}
\usage{
validation_compare_versions(
  file,
  file_mask,
  year,
  target_class_id,
  target_class_id_mask,
  version,
  multicores = 10,
  memsize = 16,
  output_dir
)
}
\arguments{
\item{file}{Character path for the classification cube file.}

\item{file_mask}{Character path for the mask cube file.}

\item{year}{Integer specifying the year to compare.}

\item{target_class_id}{Integer specifying the target class ID.}

\item{target_class_id_mask}{Integer specifying the target class ID of the mask.}

\item{version}{Character specifying the version to compare.}

\item{multicores}{Integer specifying the number of cores for parallel
processing (default: 10).}

\item{memsize}{Integer specifying the memory size for parallel
processing (default: 16).}

\item{output_dir}{Character path for the output directory.}
}
\value{
A character specifying the output file.
}
\description{
Compares the versions of a classification cube.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
