% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation_map.R
\name{vegmap_download}
\alias{vegmap_download}
\title{Download Cerrado vegetation map from S3}
\usage{
<<<<<<< HEAD
vegmap_download(year, version = "v1")
=======
vegmap_download(year)
>>>>>>> 6619145b5104f3effd2e940646afd98dff97f0ba
}
\arguments{
\item{year}{Integer or character with the year of the map (e.g. \code{2020}).}
}
\value{
Character path to the downloaded local file.
}
\description{
Download Cerrado vegetation map from S3
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
