% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi_cerrado.R
\name{roi_cerrado_biome}
\alias{roi_cerrado_biome}
\title{Get the sf object for the Cerrado biome}
\usage{
roi_cerrado_biome(
  crs = NULL,
  as_file = FALSE,
  as_convex = FALSE,
  as_union = FALSE,
  use_buffer = FALSE
)
}
\arguments{
\item{crs}{Character with the CRS to use. If not provided, the CRS of the data file is used.}

\item{as_file}{Logical with whether to return the \code{sf} object as a file path.}

\item{as_convex}{Logical with whether to return the sf object as a convex hull.}

\item{as_union}{Logical with whether to union the \code{sf} objects.}

\item{use_buffer}{Logical with whether to apply a buffer to the \code{sf} object.}
}
\value{
\code{sf} or file path object for the Cerrado biome
}
\description{
This function returns the \code{sf} object for the Cerrado biome.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}

Felipe Carvalho, \email{felipe.carvalho@inpe.br}
}
\keyword{internal}
